Wrappers rewritten to not add infinite commands

+++ olympia-v0.4.0/TeamCode/src/main/java/org/firstinspires/ftc/teamcode/framework/util/Constants.java
@@ -13,7 +13,7 @@ public class Constants {
 
     //Controller stuff
     public static final float CONTROLLER_1_DEADZONE = 0.05f;
-    public static final double CONTROLLER_2_DEADZONE = 0.05f;
+    public static final float CONTROLLER_2_DEADZONE = 0.05f;
 
     public enum ButtonID {
         A_BUTTON,
